+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Type for `Transaction.category` changed from `string` to `TransactionCategory` enum. This is a *breaking* change.
 
 ## [0.20.0] - 2019-11-18
 ### Added